+++ my-spring-test/src/main/java/org/springframework/ioc/circular/dependency/model/InstanceB.java
@@ -1,4 +1,4 @@
-package org.springframework.circular.dependency.model;
+package org.springframework.ioc.circular.dependency.model;
 
 public class InstanceB {
