+++ my-spring-test/src/main/java/org/springframework/ioc/model/User.java
@@ -1,4 +1,4 @@
-package org.springframework.model;
+package org.springframework.ioc.model;
 
 
 public class User {
